requirements.txt
streamlit>=1.30.0
Pillow>=10.0.0
gtts>=2.5.0
anthropic>=0.40.0
python-dotenv>=1.0.0
